add kernel function framework

+++ cudaSignals.h
@@ -3,37 +3,37 @@
 
 class Signals{
 	public:
-        double m_sr, m_tt; 
-   		Signals(void);    
-        Signals(double SamplingRate, double Timeinterval):m_sr{SamplingRate}, m_tt{Timeinterval}{};
+		float m_sr, m_tt; 
+		Signals(void);    
+		Signals(float SamplingRate, float Timeinterval):m_sr{SamplingRate}, m_tt{Timeinterval}{};
 		virtual bool Generator(void) = 0;
 		virtual ~Signals(void);  
 };
 
 
 class SinWave: public Signals {
-       private:
-           double m_amp, m_freq, m_ip;
-       public:
-	       SinWave(void){};
-           SinWave(double Amp, double Frequency, double InitPhase,
-                   double SamplingRate, double Timeinterval):Signals(SamplingRate, Timeinterval),
-                   m_amp{Amp}, m_freq{Frequency}, m_ip{InitPhase} {};
-	       virtual bool Generator(void);
-	       virtual ~SinWave(void){};
+	private:
+		float m_amp, m_freq, m_ip;
+	public:
+		SinWave(void){};
+		SinWave(float Amp, float Frequency, float InitPhase,
+				float SamplingRate, float Timeinterval):Signals(SamplingRate, Timeinterval),
+		m_amp{Amp}, m_freq{Frequency}, m_ip{InitPhase} {};
+		virtual bool Generator(void);
+		virtual ~SinWave(void){};
 };
 
 class CosWave: public Signals {
-       private:
-           double m_amp, m_freq, m_ip;
-       public:
-	       CosWave(void){};
-           CosWave(double Amp, double Frequency, double InitPhase,
-                   double SamplingRate, double Timeinterval):Signals(SamplingRate, Timeinterval),
-                   m_amp{Amp}, m_freq{Frequency}, m_ip{InitPhase} {};
+	private:
+		float m_amp, m_freq, m_ip;
+	public:
+		CosWave(void){};
+		CosWave(float Amp, float Frequency, float InitPhase,
+				float SamplingRate, float Timeinterval):Signals(SamplingRate, Timeinterval),
+		m_amp{Amp}, m_freq{Frequency}, m_ip{InitPhase} {};
 
-	       virtual bool Generator(void);
-	       virtual ~CosWave(void){};
+		virtual bool Generator(void);
+		virtual ~CosWave(void){};
 };
